--- a/spec/models/work/work_member_content_types_spec.rb
+++ b/spec/models/work/work_member_content_types_spec.rb
@@ -11,7 +11,38 @@ describe "Work#member_content_types" do
     ])
   end
 
-  it "finds" do
-    expect(work.member_content_types).to match(["image/tiff", "application/pdf", "audio/mpeg"])
+  describe "mode: :association" do
+    describe "pre-loaded" do
+      let(:preloaded_work) { Work.where(id: work.id).includes(members: :leaf_representative).first }
+      it "finds" do
+        expect(preloaded_work.member_content_types(mode: :association)).to match_array(["image/tiff", "application/pdf", "audio/mpeg"])
+      end
+    end
+
+    describe "none pre-loaded" do
+      let(:not_preloaded_work) { Work.where(id: work.id).first }
+
+      it "raises" do
+        expect {
+          not_preloaded_work.member_content_types(mode: :association)
+        }.to raise_error(TypeError)
+      end
+    end
+
+    describe "partially pre-loaded" do
+      let(:partially_preloaded_work) { Work.where(id: work.id).includes(:members).first }
+
+      it "raises" do
+        expect {
+          partially_preloaded_work.member_content_types(mode: :association)
+        }.to raise_error(TypeError)
+      end
+    end
+  end
+
+  describe "mode: :query" do
+    it "finds" do
+      expect(work.member_content_types(mode: :query)).to match_array(["image/tiff", "application/pdf", "audio/mpeg"])
+    end
   end
 end