When dealing with works that can have hundreds of memberes, loading all #members can be expensive. Worse is doing it more than once, or doing it in a way that causes an "n+1 query" problem, with an extra SQL query per member. We had two main problems/categories of problem discovered here:

1. To actually load the members for display, WorkShowDecorator#member_list_for_display constructs a query with SQL WHERE limits and ORDER BY clause, that does NOT end up memoized/cached in #members. So any subsequent calls to `work#members` will end up loading all members AGAIN a second time. It was expensive enough the first time, we don't want to do this. In most cases we want to replace it by a targetted SQL query getting the aggregate info we wanted -- since it's a work#show page, we can do one query per main work displayed -- there's only one main work displayed. Doing a few queries to get a few different pieces of info is okay, and better than loading #members to do it in-memory. 

2. One thing we want in a couple places is the mime types of all members. This is tricky to get! The previous implementation of work#member_content_types loaded #members and then looked at them in-memory to calculate. In some cases, if members hadn't been pre-loaded properly, this would result in an n+1 query problem. The previous implementation recognized that, and printed a warning to logs -- but we never noticed that, it was deployed to production with the warning!  The new implementation won't just warn, it will raise refusing to do it. BUT ALSO provides a different mode, where instead of looking at #members in-memory, it actually goes to database with another query. Either could be disastrously expensive depending on circumstances; the caller needs to decide which mode they want.  (Ref #895)


This is all a bit hacky, and I do worry it's easy to have a performance regression where it gets slow again when we write new code. But it's just part of how it goes I think. We also put a check in a controller spec that ensures work#members actually was NOT loaded! Can protect against some things, but not everything. 

Some attempt at measuring what effect this has on performance coming in a comment.