@jrochkind 2020-10-19
There are a variety of things/ways we can measure.  

* `time curl url >/dev/null` how long does it take curl to get it, dumping everything it gets into the trash
* Chrome inspector window time, both "time to first byte"  (TTFB) and total time
* number of object allocations in rails logs (effects RAM use, and also tends to effect performance) (just looked at Rails logs on staging vs production when I had this new code deployed on staging). 

These are all for [Ramelli](https://staging-digital.sciencehistory.org/works/4b29b614k) -- a very large/expensive page, that maybe can't exist this large anyway.  I'm just doing a few requests by hand and seeing what they look like, not being done very scientifically, so could be subject to just incidental variance. 

|   | Original | After this PR | 
| ------------- | ------------- | ------------- |
| EC2 curl /dev/null   | 2.1-2.8s  | 1.6-1.8s |
| EC2 Chrome TTFB  |  2.0s  | 1.4s | 
| EC2 Chrome total time | 3-3.5s | 1.5-2s |
| Object allocations | 1,724,882 | 1,276,613 |
| Heroku standard-1x curl /dev/null | 4.5-5.8s | 2.3-3.0s (one outlier at 3.6s) |

These measurements weren't done entirely scientifically, but these do seem to be significant improvements.